+++ quantforge/__init__.py
@@ -44,6 +44,7 @@ from .heston_mc import heston_qe_mc
 from .bates import bates_price, bates_smile
 from .double_heston import double_heston_price, double_heston_smile
 from .rough_heston import rough_heston_price, rough_heston_smile
+from .double_heston_calib import calibrate_double_heston
 from .kou import kou_price, kou_smile
 from .cgmy import cgmy_price, cgmy_smile
 from .nig import nig_price, nig_smile
@@ -192,7 +193,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "1.128.1"
+__version__ = "1.129.0"
 
 __all__ = [
     "OptionType",
@@ -245,6 +246,7 @@ __all__ = [
     "double_heston_smile",
     "rough_heston_price",
     "rough_heston_smile",
+    "calibrate_double_heston",
     "kou_price",
     "kou_smile",
     "cgmy_price",